# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Rest Clon de Airbnb",
        "version": "1.0.0",
        "description": "En esta basado en Airbnb, el objetivo es poner en practica todo el conocimiento adquirido durante el modulo de back-end con Node Js, manejando base de datos con psq y Sequelize, manejo de roles y autenticación con token JWT",
        "basePath": "v1",
        "host": "localhost:8000/api/v1"
    },
...
    "schemes": ["http","https"],
    "securityDefinitions": {
        "api_key": {
            "type": "apikey",
            "name": "api_key",
            "in": "header"
        },
        "auth": {
            "type": "jwt",
            "authorizationUrl": "/auth/login",
            "flow": "implicit"
        }
    },
    "paths": {
        "/users": {
            "get": {
                "tags": ["users"],
                "sumary": "Obtenemos todos los usuarios",
                "parameters": [],
                "produces": ["aplication/json"],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content" : {
                            "application/json": {
                                "schema" :{
                                    "type" :"object",
                                    "properties": {
                                        "id" : {
                                            "type" : "string"
                                        },
                                        "fistName": {
                                            "type": "string"
                                        },
                                        "lastName": {
                                            "type" : "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "There is no data"
                    }
                },
                "security": [{
                    "api_key": []
                }]
            }
        },
...
        "/accommodations":{
            "get":{
                "tags":["accommodations"],
                "sumary": "obtenemos todos los lugares",
                "parameters": [{
                "in":"header",
                "name":"api_key",
                "required": "true"
            }],
            "responses":{
                "200":{
                    "description": "successful operation",
                    "content" : {
                        "application/json": {
                            "schema" :{
                                "type" :"object",
                                "properties": {
                                    "id" : {
                                        "type" : "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "400":{
                    "description":"failed operation"
                }
            }
            }
        },
...
        "/reservations" : {
            "get":{
                "tags": ["reservations"],
                "sumary": "obtenemos las reservaciones",
                "parameters": [{
                    "in":"header",
                    "name":"api_key",
                    "required": "true"
                }],
                "responses":{
                    "200":{
                        "description": "successful operation",
                        "content" : {
                            "application/json": {
                                "schema" :{
                                    "type" :"object",
                                    "properties": {
                                        "id" : {
                                            "type" : "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"failed operation"
                    }
                }
            }
        }
    }
}